requirements.txt
openai>=1.0.0
openpyxl>=3.1.0
tqdm>=4.60.0
pyyaml>=6.0
streamlit>=1.30.0
pandas>=2.0.0
watchdog
